// src/models/IoK8sApiCertificatesV1CertificateSigningRequestList/types.ts
import { IoK8sApiCertificatesV1CertificateSigningRequestModelOptions as IoK8sApiCertificatesV1CertificateSigningRequest } from '../IoK8sApiCertificatesV1CertificateSigningRequest';
import { IoK8sApimachineryPkgApisMetaV1ListMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ListMeta } from '../IoK8sApimachineryPkgApisMetaV1ListMeta';

export interface IoK8sApiCertificatesV1CertificateSigningRequestListModelOptions {
  apiVersion?: string;
  items: Array<IoK8sApiCertificatesV1CertificateSigningRequest>;
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ListMeta;
}